+++ pkg.nix
@@ -3,11 +3,11 @@
 }:
 mkDerivation {
   pname = "flay";
-  version = "0.4";
+  version = "0.5";
   src = ./.;
   libraryHaskellDepends = [ base constraints transformers ];
   testHaskellDepends = [ base tasty tasty-quickcheck transformers ];
   homepage = "https://github.com/k0001/flay";
-  description = "Work generically on your datatype without knowing its shape nor its contents";
+  description = "Generic programming for higher-kinded types";
   license = stdenv.lib.licenses.bsd3;
 }